got order process working, just need to fix the payment part to complete full order flow

+++ services/restaurant-server/src/main/java/com/devshaks/delivery/cuisine/CuisineRequest.java
@@ -3,11 +3,13 @@ package com.devshaks.delivery.cuisine;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record CuisineRequest(
         Integer id,
         @NotBlank(message = "Cuisine Name is Required") String name,
         @NotBlank(message = "Cuisine Description is Required") String description,
-        @NotNull(message = "Cuisine Price is Required") Double price
+        @NotNull(message = "Cuisine Price is Required") BigDecimal price
 ) {
 
 }
